+++ openpyxl2/worksheet/tests/test_table.py
@@ -45,14 +45,17 @@ def Table():
 class TestTable:
 
     def test_ctor(self, Table, TableColumn):
-        table = Table(displayName="A_Sample_Table", ref="A1:F10")
-        table.tableColumns.append(TableColumn(id=1, name="Column1"))
+        table = Table(displayName="A_Sample_Table", ref="A1:D5")
         xml = tostring(table.to_tree())
         expected = """
         <table xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main"
-           displayName="A_Sample_Table" name="A_Sample_Table" id="1" ref="A1:F10">
-        <tableColumns count="1">
+           displayName="A_Sample_Table" name="A_Sample_Table" id="1" ref="A1:D5">
+           <autoFilter ref="A1:D5" />
+        <tableColumns count="4">
           <tableColumn id="1" name="Column1" />
+          <tableColumn id="2" name="Column2" />
+          <tableColumn id="3" name="Column3" />
+          <tableColumn id="4" name="Column4" />
         </tableColumns>
         </table>
         """
@@ -72,14 +75,9 @@ class TestTable:
                               ref="A1:AA27")
 
 
-    def test_local_path(self, Table):
+    def test_path(self, Table):
         table = Table(displayName="Table1", ref="A1:M6")
-        assert table.path == "/tables/table1.xml"
-
-
-    def test_archive_path(self, Table):
-        table = Table(displayName="Table1", ref="A1:M6")
-        assert table.abs_path == "/xl/tables/table1.xml"
+        assert table.path == "/xl/tables/table1.xml"
 
 
     def test_write(self, Table):
@@ -87,7 +85,7 @@ class TestTable:
         archive = ZipFile(out, "w")
         table = Table(displayName="Table1", ref="B1:L10")
         table._write(archive)
-        assert "/xl/tables/table1.xml" in archive.namelist()
+        assert "xl/tables/table1.xml" in archive.namelist()
 
 
 @pytest.fixture