+++ openpyxl2/comments/shape_writer.py
@@ -61,7 +61,8 @@ class ShapeWriter(object):
 
 
     def write(self, root):
-        if root is None:
+
+        if not hasattr(root, "findall"):
             root = Element("xml")
         # Remove any existing comment shapes
         comments = root.findall("{%s}shape[@type='#_x0000_t202']" % vmlns)