+++ doc/changes.rst
@@ -24,6 +24,7 @@ Bug fixes
 
 * `#625 <https://bitbucket.org/openpyxl/openpyxl/issues/625>`_ Exception
 raises when inspecting EmptyCells in read-only mode
+* `547`<https://bitbucket.org/openpyxl/openpyxl/issues/547>`_ Functions for handling OOXML "escaped" ST_XStrings
 
 
 2.4.0-a1 (2016-04-11)